src/pyproject.toml
[tool.poetry]
name = "equipment-api"
version = "0.1.0"
description = "тестовое задание"
authors = ["azat715 <azat715@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "equipment_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
pytest = "^7.1.3"
mysqlclient = "^2.1.1"
django-environ = "^0.9.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
